move more code from bin to lib
